refactor(tests): rename project_scopes_repository to project_scope_repository
- update tests/integration/repositories/mod.rs to use the new module name
- rename the file accordingly
