+++ tests/integration/repositories/mod.rs
@@ -1,3 +1,3 @@
 pub mod project_repository;
-pub mod project_scopes_repository;
+pub mod project_scope_repository;
 pub mod service_account_repository;
